move method for creating TypeNode to appropriate module
