+++ src/config.js
@@ -3,4 +3,6 @@ const customTrackingHeader = {
   value: 'gatsby-source-kentico-cloud;3.0.0-beta1',
 };
 
-exports.customTrackingHeader = customTrackingHeader;
+module.exports = {
+  customTrackingHeader,
+};
